refactor!: migrate to multi-module architecture
BREAKING CHANGE: Project restructured into separate modules

- feat(plugin): Move plugin code to symbolcraft-plugin module
- feat(runtime): Add new symbolcraft-runtime module
- feat(material-symbols): Add symbolcraft-material-symbols module
