+++ .github/workflows/material-symbols-ci.yml
@@ -1,0 +1,100 @@
+name: Material Symbols CI
+
+on:
+  push:
+    branches: [ "**" ]
+    paths:
+      - 'symbolcraft-material-symbols/**'
+      - 'symbolcraft-runtime/**'  # Depends on runtime
+      - 'gradle/libs.versions.toml'
+      - 'gradle.properties'
+      - 'build.gradle.kts'
+      - 'settings.gradle.kts'
+      - '.github/workflows/material-symbols-ci.yml'
+  pull_request:
+    branches: [ "**" ]
+    paths:
+      - 'symbolcraft-material-symbols/**'
+      - 'symbolcraft-runtime/**'
+      - 'gradle/libs.versions.toml'
+      - 'gradle.properties'
+      - 'build.gradle.kts'
+      - 'settings.gradle.kts'
+
+env:
+  GRADLE_OPTS: -Dorg.gradle.daemon=false -Dorg.gradle.parallel=true -Dorg.gradle.caching=true
+
+jobs:
+  build:
+    name: Build Material Symbols
+    runs-on: macos-latest  # macOS for iOS builds
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v5
+
+    - name: Set up JDK 17
+      uses: actions/setup-java@v5
+      with:
+        java-version: '17'
+        distribution: 'corretto'
+
+    - name: Setup Gradle
+      uses: gradle/actions/setup-gradle@v4
+
+    - name: Verify Kotlin version
+      run: |
+        echo "Checking Kotlin version..."
+        ./gradlew :symbolcraft-material-symbols:dependencies --configuration runtimeClasspath | grep kotlin-stdlib || true
+        echo "Expected Kotlin version: 2.0.0"
+
+    - name: Build material-symbols (all platforms)
+      run: ./gradlew :symbolcraft-material-symbols:build -x test --stacktrace --no-configuration-cache
+
+    - name: Run tests
+      run: ./gradlew :symbolcraft-material-symbols:test --stacktrace
+
+    - name: Upload Android AAR
+      uses: actions/upload-artifact@v4
+      with:
+        name: material-symbols-aar
+        path: symbolcraft-material-symbols/build/outputs/aar/*.aar
+        retention-days: 7
+
+    - name: Upload iOS Framework
+      uses: actions/upload-artifact@v4
+      with:
+        name: material-symbols-ios-framework
+        path: symbolcraft-material-symbols/build/XCFrameworks/release/*.xcframework
+        retention-days: 7
+
+  validate:
+    name: Validate Material Symbols
+    runs-on: macos-latest
+    needs: build
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v5
+
+    - name: Set up JDK 17
+      uses: actions/setup-java@v5
+      with:
+        java-version: '17'
+        distribution: 'corretto'
+
+    - name: Setup Gradle
+      uses: gradle/actions/setup-gradle@v4
+
+    - name: Publish to local Maven
+      run: ./gradlew :symbolcraft-material-symbols:publishToMavenLocal --stacktrace
+
+    - name: Verify publication
+      run: |
+        if [ -d ~/.m2/repository/io/github/kingsword09/symbolcraft-material-symbols ]; then
+          echo "✅ Material Symbols published to local Maven"
+          find ~/.m2/repository/io/github/kingsword09/symbolcraft-material-symbols -type f | head -10
+        else
+          echo "❌ Material Symbols not found in local Maven"
+          exit 1
+        fi
