+++ .github/workflows/plugin-ci.yml
@@ -1,0 +1,85 @@
+name: Plugin CI
+
+on:
+  push:
+    branches: [ "**" ]
+    paths:
+      - 'symbolcraft-plugin/**'
+      - 'gradle/libs.versions.toml'
+      - 'gradle.properties'
+      - 'build.gradle.kts'
+      - 'settings.gradle.kts'
+      - '.github/workflows/plugin-ci.yml'
+  pull_request:
+    branches: [ "**" ]
+    paths:
+      - 'symbolcraft-plugin/**'
+      - 'gradle/libs.versions.toml'
+      - 'gradle.properties'
+      - 'build.gradle.kts'
+      - 'settings.gradle.kts'
+
+env:
+  GRADLE_OPTS: -Dorg.gradle.daemon=false -Dorg.gradle.parallel=true -Dorg.gradle.caching=true
+
+jobs:
+  build:
+    name: Build Plugin
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v5
+
+    - name: Set up JDK 17
+      uses: actions/setup-java@v5
+      with:
+        java-version: '17'
+        distribution: 'corretto'
+
+    - name: Setup Gradle
+      uses: gradle/actions/setup-gradle@v4
+
+    - name: Build plugin
+      run: ./gradlew :symbolcraft-plugin:build --stacktrace
+
+    - name: Run plugin tests
+      run: ./gradlew :symbolcraft-plugin:test --stacktrace
+
+    - name: Upload build artifacts
+      uses: actions/upload-artifact@v4
+      with:
+        name: plugin-artifacts
+        path: symbolcraft-plugin/build/libs/*.jar
+        retention-days: 7
+
+  validate:
+    name: Validate Plugin
+    runs-on: ubuntu-latest
+    needs: build
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v5
+
+    - name: Set up JDK 17
+      uses: actions/setup-java@v5
+      with:
+        java-version: '17'
+        distribution: 'corretto'
+
+    - name: Setup Gradle
+      uses: gradle/actions/setup-gradle@v4
+
+    - name: Publish to local Maven
+      run: ./gradlew :symbolcraft-plugin:publishToMavenLocal --stacktrace
+
+    - name: Verify publication
+      run: |
+        if [ -d ~/.m2/repository/io/github/kingsword09/symbolcraft ]; then
+          echo "✅ Plugin published to local Maven"
+          find ~/.m2/repository/io/github/kingsword09/symbolcraft -name "*.jar" | head -5
+        else
+          echo "❌ Plugin not found in local Maven"
+          exit 1
+        fi
