+++ .github/workflows/release-all.yml
@@ -1,0 +1,197 @@
+name: Release All Modules
+
+on:
+  push:
+    branches: [ main ]
+
+env:
+  GRADLE_OPTS: -Dorg.gradle.daemon=false -Dorg.gradle.parallel=true -Dorg.gradle.caching=true
+
+jobs:
+  check-commit:
+    name: Check Commit Message
+    runs-on: ubuntu-latest
+    outputs:
+      should_release: ${{ steps.check.outputs.should_release }}
+      version: ${{ steps.extract.outputs.version }}
+    
+    steps:
+    - name: Check commit message
+      id: check
+      run: |
+        COMMIT_MSG="${{ github.event.head_commit.message }}"
+        if [[ "$COMMIT_MSG" =~ ^chore\(release\): ]] || [[ "$COMMIT_MSG" =~ ^chore\(all\): ]]; then
+          echo "should_release=true" >> $GITHUB_OUTPUT
+        else
+          echo "should_release=false" >> $GITHUB_OUTPUT
+        fi
+
+    - name: Extract version
+      id: extract
+      if: steps.check.outputs.should_release == 'true'
+      run: |
+        COMMIT_MSG="${{ github.event.head_commit.message }}"
+        VERSION=$(echo "$COMMIT_MSG" | grep -oP 'chore\((release|all)\):\s*v?\K[0-9]+\.[0-9]+\.[0-9]+' || echo "")
+        if [ -z "$VERSION" ]; then
+          echo "❌ No version found in commit message"
+          echo "Expected format: chore(release): v1.2.3 or chore(all): 1.2.3"
+          exit 1
+        fi
+        echo "version=$VERSION" >> $GITHUB_OUTPUT
+        echo "✅ Extracted version: $VERSION"
+
+  release:
+    name: Build and Release All Modules
+    runs-on: macos-latest  # macOS for iOS builds
+    needs: check-commit
+    if: needs.check-commit.outputs.should_release == 'true'
+    permissions:
+      contents: write
+      packages: write
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v5
+      with:
+        fetch-depth: 0
+
+    - name: Set up JDK 17
+      uses: actions/setup-java@v5
+      with:
+        java-version: '17'
+        distribution: 'corretto'
+
+    - name: Setup Gradle
+      uses: gradle/actions/setup-gradle@v4
+
+    - name: Update version
+      run: |
+        VERSION=${{ needs.check-commit.outputs.version }}
+        sed -i '' "s/version = \".*\"/version = \"$VERSION\"/" build.gradle.kts
+        echo "✅ Updated version to $VERSION"
+        grep "version =" build.gradle.kts
+
+    - name: Build all modules
+      run: ./gradlew clean build -x test --stacktrace
+
+    - name: Create GitHub Release
+      uses: softprops/action-gh-release@v2
+      with:
+        tag_name: v${{ needs.check-commit.outputs.version }}
+        name: SymbolCraft v${{ needs.check-commit.outputs.version }}
+        body: |
+          ## SymbolCraft v${{ needs.check-commit.outputs.version }}
+          
+          Complete release including all modules.
+          
+          ### 📦 Modules
+          
+          **Plugin**
+          ```kotlin
+          plugins {
+              id("io.github.kingsword09.symbolcraft") version "${{ needs.check-commit.outputs.version }}"
+          }
+          ```
+          
+          **Runtime**
+          ```kotlin
+          dependencies {
+              implementation("io.github.kingsword09:symbolcraft-runtime:${{ needs.check-commit.outputs.version }}")
+          }
+          ```
+          
+          **Material Symbols**
+          ```kotlin
+          dependencies {
+              implementation("io.github.kingsword09:symbolcraft-material-symbols:${{ needs.check-commit.outputs.version }}")
+          }
+          ```
+          
+          ### 🔗 Links
+          - [Documentation](https://github.com/kingsword09/SymbolCraft#readme)
+          - [Gradle Plugin Portal](https://plugins.gradle.org/plugin/io.github.kingsword09.symbolcraft)
+          - [Maven Central](https://central.sonatype.com/namespace/io.github.kingsword09)
+        files: |
+          symbolcraft-plugin/build/libs/*.jar
+          symbolcraft-runtime/build/outputs/aar/*.aar
+          symbolcraft-runtime/build/libs/*.jar
+          symbolcraft-material-symbols/build/outputs/aar/*.aar
+          symbolcraft-material-symbols/build/libs/*.jar
+        draft: false
+        prerelease: false
+        generate_release_notes: true
+
+  publish-all:
+    name: Publish All to Maven Central
+    runs-on: macos-latest
+    needs: [check-commit, release]
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v5
+
+    - name: Set up JDK 17
+      uses: actions/setup-java@v5
+      with:
+        java-version: '17'
+        distribution: 'corretto'
+
+    - name: Setup Gradle
+      uses: gradle/actions/setup-gradle@v4
+
+    - name: Update version
+      run: |
+        VERSION=${{ needs.check-commit.outputs.version }}
+        sed -i '' "s/version = \".*\"/version = \"$VERSION\"/" build.gradle.kts
+
+    - name: Publish all modules to Maven Central
+      env:
+        ORG_GRADLE_PROJECT_mavenCentralUsername: ${{ secrets.OSSRH_USERNAME }}
+        ORG_GRADLE_PROJECT_mavenCentralPassword: ${{ secrets.OSSRH_PASSWORD }}
+        ORG_GRADLE_PROJECT_signingInMemoryKey: ${{ secrets.SIGNING_KEY }}
+        ORG_GRADLE_PROJECT_signingInMemoryKeyPassword: ${{ secrets.SIGNING_PASSWORD }}
+      run: ./gradlew publishAll --no-configuration-cache --stacktrace
+
+  publish-gradle-portal:
+    name: Publish Plugin to Gradle Portal
+    runs-on: ubuntu-latest
+    needs: [check-commit, release]
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v5
+
+    - name: Set up JDK 17
+      uses: actions/setup-java@v5
+      with:
+        java-version: '17'
+        distribution: 'corretto'
+
+    - name: Setup Gradle
+      uses: gradle/actions/setup-gradle@v4
+
+    - name: Update version
+      run: |
+        VERSION=${{ needs.check-commit.outputs.version }}
+        sed -i "s/version = \".*\"/version = \"$VERSION\"/" build.gradle.kts
+
+    - name: Publish plugin to Gradle Plugin Portal
+      env:
+        GRADLE_PUBLISH_KEY: ${{ secrets.GRADLE_PUBLISH_KEY }}
+        GRADLE_PUBLISH_SECRET: ${{ secrets.GRADLE_PUBLISH_SECRET }}
+      run: ./gradlew :symbolcraft-plugin:publishPlugins --stacktrace
+
+  notify:
+    name: Notify Success
+    runs-on: ubuntu-latest
+    needs: [publish-all, publish-gradle-portal]
+    if: success()
+
+    steps:
+    - name: Success notification
+      run: |
+        echo "🎉 SymbolCraft v${{ needs.check-commit.outputs.version }} released!"
+        echo "✅ All modules published to Maven Central"
+        echo "✅ Plugin published to Gradle Plugin Portal"
+        echo "🔗 Plugin: https://plugins.gradle.org/plugin/io.github.kingsword09.symbolcraft"
+        echo "🔗 Maven: https://central.sonatype.com/namespace/io.github.kingsword09"
