+++ .github/workflows/release-material-symbols.yml
@@ -1,0 +1,151 @@
+name: Release Material Symbols
+
+on:
+  push:
+    branches: [ main ]
+    paths:
+      - 'symbolcraft-material-symbols/**'
+
+env:
+  GRADLE_OPTS: -Dorg.gradle.daemon=false -Dorg.gradle.parallel=true -Dorg.gradle.caching=true
+
+jobs:
+  check-commit:
+    name: Check Commit Message
+    runs-on: ubuntu-latest
+    outputs:
+      should_release: ${{ steps.check.outputs.should_release }}
+      version: ${{ steps.extract.outputs.version }}
+    
+    steps:
+    - name: Check commit message
+      id: check
+      run: |
+        COMMIT_MSG="${{ github.event.head_commit.message }}"
+        if [[ "$COMMIT_MSG" =~ ^chore\(material-symbols\): ]]; then
+          echo "should_release=true" >> $GITHUB_OUTPUT
+        else
+          echo "should_release=false" >> $GITHUB_OUTPUT
+        fi
+
+    - name: Extract version
+      id: extract
+      if: steps.check.outputs.should_release == 'true'
+      run: |
+        COMMIT_MSG="${{ github.event.head_commit.message }}"
+        VERSION=$(echo "$COMMIT_MSG" | grep -oP 'chore\(material-symbols\):\s*v?\K[0-9]+\.[0-9]+\.[0-9]+' || echo "")
+        if [ -z "$VERSION" ]; then
+          echo "❌ No version found in commit message"
+          echo "Expected format: chore(material-symbols): v1.2.3 or chore(material-symbols): 1.2.3"
+          exit 1
+        fi
+        echo "version=$VERSION" >> $GITHUB_OUTPUT
+        echo "✅ Extracted version: $VERSION"
+
+  release:
+    name: Build and Release Material Symbols
+    runs-on: macos-latest  # macOS for iOS builds
+    needs: check-commit
+    if: needs.check-commit.outputs.should_release == 'true'
+    permissions:
+      contents: write
+      packages: write
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v5
+      with:
+        fetch-depth: 0
+
+    - name: Set up JDK 17
+      uses: actions/setup-java@v5
+      with:
+        java-version: '17'
+        distribution: 'corretto'
+
+    - name: Setup Gradle
+      uses: gradle/actions/setup-gradle@v4
+
+    - name: Update version
+      run: |
+        VERSION=${{ needs.check-commit.outputs.version }}
+        sed -i '' "s/version = \".*\"/version = \"$VERSION\"/" build.gradle.kts
+        echo "✅ Updated version to $VERSION"
+
+    - name: Build material-symbols (all platforms)
+      run: ./gradlew :symbolcraft-material-symbols:build -x test --stacktrace
+
+    - name: Create GitHub Release
+      uses: softprops/action-gh-release@v2
+      with:
+        tag_name: material-symbols-v${{ needs.check-commit.outputs.version }}
+        name: Material Symbols v${{ needs.check-commit.outputs.version }}
+        body: |
+          ## SymbolCraft Material Symbols v${{ needs.check-commit.outputs.version }}
+          
+          Pre-generated Material Symbols icons for Compose Multiplatform.
+          
+          ### 📦 Installation
+          ```kotlin
+          dependencies {
+              implementation("io.github.kingsword09:symbolcraft-material-symbols:${{ needs.check-commit.outputs.version }}")
+          }
+          ```
+          
+          ### 🎯 Platforms
+          - Android (API 21+)
+          - JVM (Java 17+)
+          - iOS (arm64, x64, simulatorArm64)
+          
+          ### 🔗 Links
+          - [Documentation](https://github.com/kingsword09/SymbolCraft/blob/main/symbolcraft-material-symbols/README.md)
+        files: |
+          symbolcraft-material-symbols/build/outputs/aar/*.aar
+          symbolcraft-material-symbols/build/libs/*.jar
+        draft: false
+        prerelease: false
+        generate_release_notes: true
+
+  publish-maven-central:
+    name: Publish to Maven Central
+    runs-on: macos-latest
+    needs: [check-commit, release]
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v5
+
+    - name: Set up JDK 17
+      uses: actions/setup-java@v5
+      with:
+        java-version: '17'
+        distribution: 'corretto'
+
+    - name: Setup Gradle
+      uses: gradle/actions/setup-gradle@v4
+
+    - name: Update version
+      run: |
+        VERSION=${{ needs.check-commit.outputs.version }}
+        sed -i '' "s/version = \".*\"/version = \"$VERSION\"/" build.gradle.kts
+
+    - name: Publish to Maven Central
+      env:
+        ORG_GRADLE_PROJECT_mavenCentralUsername: ${{ secrets.OSSRH_USERNAME }}
+        ORG_GRADLE_PROJECT_mavenCentralPassword: ${{ secrets.OSSRH_PASSWORD }}
+        ORG_GRADLE_PROJECT_signingInMemoryKey: ${{ secrets.SIGNING_KEY }}
+        ORG_GRADLE_PROJECT_signingInMemoryKeyPassword: ${{ secrets.SIGNING_PASSWORD }}
+      run: ./gradlew :symbolcraft-material-symbols:publishToMavenCentral --no-configuration-cache --stacktrace
+
+  notify:
+    name: Notify Success
+    runs-on: ubuntu-latest
+    needs: [publish-maven-central]
+    if: success()
+
+    steps:
+    - name: Success notification
+      run: |
+        echo "🎉 Material Symbols v${{ needs.check-commit.outputs.version }} released!"
+        echo "✅ Published to Maven Central"
+        echo "🔗 https://central.sonatype.com/artifact/io.github.kingsword09/symbolcraft-material-symbols"
