+++ .github/workflows/release-plugin.yml
@@ -1,0 +1,177 @@
+name: Release Plugin
+
+on:
+  push:
+    branches: [ main ]
+    paths:
+      - 'symbolcraft-plugin/**'
+
+env:
+  GRADLE_OPTS: -Dorg.gradle.daemon=false -Dorg.gradle.parallel=true -Dorg.gradle.caching=true
+
+jobs:
+  check-commit:
+    name: Check Commit Message
+    runs-on: ubuntu-latest
+    outputs:
+      should_release: ${{ steps.check.outputs.should_release }}
+      version: ${{ steps.extract.outputs.version }}
+    
+    steps:
+    - name: Check commit message
+      id: check
+      run: |
+        COMMIT_MSG="${{ github.event.head_commit.message }}"
+        if [[ "$COMMIT_MSG" =~ ^chore\(plugin\): ]]; then
+          echo "should_release=true" >> $GITHUB_OUTPUT
+        else
+          echo "should_release=false" >> $GITHUB_OUTPUT
+        fi
+
+    - name: Extract version
+      id: extract
+      if: steps.check.outputs.should_release == 'true'
+      run: |
+        COMMIT_MSG="${{ github.event.head_commit.message }}"
+        VERSION=$(echo "$COMMIT_MSG" | grep -oP 'chore\(plugin\):\s*v?\K[0-9]+\.[0-9]+\.[0-9]+' || echo "")
+        if [ -z "$VERSION" ]; then
+          echo "❌ No version found in commit message"
+          echo "Expected format: chore(plugin): v1.2.3 or chore(plugin): 1.2.3"
+          exit 1
+        fi
+        echo "version=$VERSION" >> $GITHUB_OUTPUT
+        echo "✅ Extracted version: $VERSION"
+
+  release:
+    name: Build and Release Plugin
+    runs-on: ubuntu-latest
+    needs: check-commit
+    if: needs.check-commit.outputs.should_release == 'true'
+    permissions:
+      contents: write
+      packages: write
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v5
+      with:
+        fetch-depth: 0
+
+    - name: Set up JDK 17
+      uses: actions/setup-java@v5
+      with:
+        java-version: '17'
+        distribution: 'corretto'
+
+    - name: Setup Gradle
+      uses: gradle/actions/setup-gradle@v4
+
+    - name: Update version
+      run: |
+        VERSION=${{ needs.check-commit.outputs.version }}
+        # Update root build.gradle.kts
+        sed -i "s/version = \".*\"/version = \"$VERSION\"/" build.gradle.kts
+        echo "✅ Updated version to $VERSION"
+        grep "version =" build.gradle.kts
+
+    - name: Build plugin
+      run: ./gradlew :symbolcraft-plugin:build --stacktrace
+
+    - name: Create GitHub Release
+      uses: softprops/action-gh-release@v2
+      with:
+        tag_name: plugin-v${{ needs.check-commit.outputs.version }}
+        name: Plugin v${{ needs.check-commit.outputs.version }}
+        body: |
+          ## SymbolCraft Plugin v${{ needs.check-commit.outputs.version }}
+          
+          Gradle plugin for generating icons on-demand from multiple libraries.
+          
+          ### 📦 Installation
+          ```kotlin
+          plugins {
+              id("io.github.kingsword09.symbolcraft") version "${{ needs.check-commit.outputs.version }}"
+          }
+          ```
+          
+          ### 🔗 Links
+          - [Documentation](https://github.com/kingsword09/SymbolCraft/blob/main/symbolcraft-plugin/README.md)
+          - [Gradle Plugin Portal](https://plugins.gradle.org/plugin/io.github.kingsword09.symbolcraft)
+        files: symbolcraft-plugin/build/libs/*.jar
+        draft: false
+        prerelease: false
+        generate_release_notes: true
+
+  publish-gradle-portal:
+    name: Publish to Gradle Plugin Portal
+    runs-on: ubuntu-latest
+    needs: [check-commit, release]
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v5
+
+    - name: Set up JDK 17
+      uses: actions/setup-java@v5
+      with:
+        java-version: '17'
+        distribution: 'corretto'
+
+    - name: Setup Gradle
+      uses: gradle/actions/setup-gradle@v4
+
+    - name: Update version
+      run: |
+        VERSION=${{ needs.check-commit.outputs.version }}
+        sed -i "s/version = \".*\"/version = \"$VERSION\"/" build.gradle.kts
+
+    - name: Publish to Gradle Plugin Portal
+      env:
+        GRADLE_PUBLISH_KEY: ${{ secrets.GRADLE_PUBLISH_KEY }}
+        GRADLE_PUBLISH_SECRET: ${{ secrets.GRADLE_PUBLISH_SECRET }}
+      run: ./gradlew :symbolcraft-plugin:publishPlugins --stacktrace
+
+  publish-maven-central:
+    name: Publish to Maven Central
+    runs-on: ubuntu-latest
+    needs: [check-commit, release]
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v5
+
+    - name: Set up JDK 17
+      uses: actions/setup-java@v5
+      with:
+        java-version: '17'
+        distribution: 'corretto'
+
+    - name: Setup Gradle
+      uses: gradle/actions/setup-gradle@v4
+
+    - name: Update version
+      run: |
+        VERSION=${{ needs.check-commit.outputs.version }}
+        sed -i "s/version = \".*\"/version = \"$VERSION\"/" build.gradle.kts
+
+    - name: Publish to Maven Central
+      env:
+        ORG_GRADLE_PROJECT_mavenCentralUsername: ${{ secrets.OSSRH_USERNAME }}
+        ORG_GRADLE_PROJECT_mavenCentralPassword: ${{ secrets.OSSRH_PASSWORD }}
+        ORG_GRADLE_PROJECT_signingInMemoryKey: ${{ secrets.SIGNING_KEY }}
+        ORG_GRADLE_PROJECT_signingInMemoryKeyPassword: ${{ secrets.SIGNING_PASSWORD }}
+      run: ./gradlew :symbolcraft-plugin:publishToMavenCentral --no-configuration-cache --stacktrace
+
+  notify:
+    name: Notify Success
+    runs-on: ubuntu-latest
+    needs: [publish-gradle-portal, publish-maven-central]
+    if: success()
+
+    steps:
+    - name: Success notification
+      run: |
+        echo "🎉 Plugin v${{ needs.check-commit.outputs.version }} released!"
+        echo "✅ Published to Gradle Plugin Portal"
+        echo "✅ Published to Maven Central"
+        echo "🔗 https://plugins.gradle.org/plugin/io.github.kingsword09.symbolcraft"
