+++ example/gradle.properties
@@ -1,6 +1,7 @@
 #Kotlin
 kotlin.code.style=official
 kotlin.daemon.jvmargs=-Xmx3072M
+kotlin.mpp.androidGradlePluginCompatibility.nowarn=true
 
 #Gradle
 org.gradle.jvmargs=-Xmx4096M -Dfile.encoding=UTF-8
@@ -10,3 +11,11 @@ org.gradle.caching=true
 #Android
 android.nonTransitiveRClass=true
 android.useAndroidX=true
+
+#Compose
+org.jetbrains.compose.experimental.uikit.enabled=true
+org.jetbrains.compose.reload.isEnabled=true
+org.jetbrains.compose.reload.quiet=true
+
+# Disable unsupported Compose Compiler options
+compose.kotlinCompilerPluginGenerateFunctionKeyMetaClasses=false
