+++ example/gradle/libs.versions.toml
@@ -1,21 +1,21 @@
 [versions]
-agp = "8.11.2"
-android-compileSdk = "36"
+agp = "8.7.3"
+android-compileSdk = "35"
 android-minSdk = "24"
-android-targetSdk = "36"
-androidx-activity = "1.11.0"
-androidx-appcompat = "1.7.1"
-androidx-core = "1.17.0"
-androidx-espresso = "3.7.0"
-androidx-lifecycle = "2.9.4"
-androidx-testExt = "1.3.0"
+android-targetSdk = "35"
+androidx-activity = "1.9.2"
+androidx-appcompat = "1.7.0"
+androidx-core = "1.13.1"
+androidx-espresso = "3.6.1"
+androidx-lifecycle = "2.8.0"
+androidx-testExt = "1.2.1"
 composeHotReload = "1.0.0-beta07"
-composeMultiplatform = "1.9.0"
+composeMultiplatform = "1.7.0"
 junit = "4.13.2"
-kotlin = "2.2.20"
-kotlinx-coroutines = "1.10.2"
+kotlin = "2.0.0"
+kotlinx-coroutines = "1.8.1"
 
-symbolcraft = "0.3.2"
+symbolcraft = "0.4.0"
 
 [libraries]
 kotlin-test = { module = "org.jetbrains.kotlin:kotlin-test", version.ref = "kotlin" }
