+++ gradle.properties
@@ -4,7 +4,15 @@ org.gradle.caching=true
 
 # Kotlin
 kotlin.code.style=official
+kotlin.mpp.stability.nowarn=true
+kotlin.mpp.androidSourceSetLayoutVersion=2
+kotlin.mpp.androidGradlePluginCompatibility.nowarn=true
+kotlin.apple.xcodeCompatibility.nowarn=true
+
+# AndroidX configuration
+android.useAndroidX=true
+android.enableJetifier=false
 
 # Dokka V2
 org.jetbrains.dokka.experimental.gradle.pluginMode=V2Enabled
-#org.jetbrains.dokka.experimental.gradle.pluginMode.noWarn=true
+org.jetbrains.dokka.experimental.gradle.pluginMode.noWarn=true
