+++ settings.gradle.kts
@@ -4,7 +4,12 @@ enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 pluginManagement {
     repositories {
         gradlePluginPortal()
-        google()
+        google {
+            mavenContent {
+                includeGroupByRegex(".*google.*")
+                includeGroupByRegex(".*android.*")
+            }
+        }
         mavenCentral()
     }
 }
@@ -12,9 +17,23 @@ pluginManagement {
 dependencyResolutionManagement {
     repositoriesMode.set(RepositoriesMode.PREFER_SETTINGS)
     repositories {
-        google()
+        google {
+            mavenContent {
+                includeGroupByRegex(".*android.*")
+                includeGroupByRegex(".*androidx.*")
+                includeGroupByRegex(".*google.*")
+            }
+        }
         mavenCentral()
+        gradlePluginPortal()
     }
 }
 
 rootProject.name = "SymbolCraft"
+
+// Include all submodules
+include(":symbolcraft-plugin")
+include(":symbolcraft-runtime")
+include(":symbolcraft-material-symbols")
+
+// Note: example 作为独立项目，避免插件版本冲突，build it separately in example/ directory
