+++ symbolcraft-material-symbols/src/commonMain/kotlin/io/github/kingsword09/symbolcraft/materialsymbols/MaterialSymbols.kt
@@ -1,0 +1,24 @@
+package io.github.kingsword09.symbolcraft.materialsymbols
+
+/**
+ * MaterialSymbols - Pre-generated Material Symbols icons
+ *
+ * This module will provide:
+ * - Pre-generated Material Symbols icons in all variants (Outlined, Rounded, Sharp)
+ * - Multiple font weights (100-700)
+ * - Fill states (0 = unfilled, 1 = filled)
+ * - Lazy loading via runtime module
+ *
+ * Status: Under Development
+ */
+object MaterialSymbols {
+    /**
+     * Library version
+     */
+    const val VERSION = "0.4.0"
+    
+    /**
+     * Total number of icons (placeholder)
+     */
+    const val TOTAL_ICONS = 0
+}
