+++ symbolcraft-plugin/src/main/kotlin/io/github/kingsword09/symbolcraft/converter/Svg2ComposeConverter.kt
