+++ symbolcraft-plugin/src/main/kotlin/io/github/kingsword09/symbolcraft/plugin/samples/LocalIconsSamples.kt
