+++ symbolcraft-plugin/src/main/kotlin/io/github/kingsword09/symbolcraft/tasks/ValidateSymbolsConfigTask.kt
