+++ symbolcraft-plugin/src/test/kotlin/io/github/kingsword09/symbolcraft/converter/IconNameTransformerTest.kt
