+++ symbolcraft-plugin/src/test/kotlin/io/github/kingsword09/symbolcraft/plugin/LocalIconsBuilderTest.kt
