+++ symbolcraft-plugin/src/test/kotlin/io/github/kingsword09/symbolcraft/tasks/GenerateSymbolsTaskTest.kt
