+++ symbolcraft-runtime/src/commonMain/kotlin/io/github/kingsword09/symbolcraft/runtime/SymbolCraftRuntime.kt
@@ -1,0 +1,23 @@
+package io.github.kingsword09.symbolcraft.runtime
+
+/**
+ * SymbolCraft Runtime Library
+ *
+ * This module will provide runtime support for:
+ * - Lazy icon loading
+ * - Memory caching
+ * - Multi-platform icon access
+ *
+ * Status: Under Development
+ */
+object SymbolCraftRuntime {
+    /**
+     * Library version
+     */
+    const val VERSION = "0.4.0"
+    
+    /**
+     * Placeholder function - will be implemented in future iterations
+     */
+    fun isInitialized(): Boolean = true
+}
